+++ SimNextgenApp.Demo/RestaurantSample/CustomerGroup.cs
@@ -1,10 +1,9 @@
 ﻿namespace SimNextgenApp.Demo.RestaurantSample;
 
-internal class CustomerGroup(int size, double arrivalTime)
+internal class CustomerGroup(int size, long arrivalTime)
 {
     public int Id { get; } = Interlocked.Increment(ref _nextId);
     public int GroupSize { get; } = size;
-    public double ArrivalTime { get; set; } = arrivalTime;
-
+    public long ArrivalTime { get; set; } = arrivalTime;
     private static int _nextId = 0;
 }
